requirements.txt
Django>=5.0.6
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pyotp>=2.9.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
PyWavelets>=1.0.0  # For wavelet transforms used in steganography
